requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
python-dotenv